\name{BEF.dat}
\docType{data}

\alias{BEF.dat}


\title{Bartlett Experimental Forest inventory data}

\description{
  Data generated in long-term research studies on the Bartlett
  Experimental Forest, Bartlett, NH funded by the U.S. Department
  of Agriculture, Forest Service, Northeastern Research Station.

  This dataset holds 1991 and 2002 forest inventory data for 437 points
  on the BEF.dat. Variables include species specific basal area and biomass;
  inventory plot coordinates; slope; elevation; and tasseled cap brightness (TC1),
  greenness (TC2), and wetness (TC3) components from spring, summer,
  and fall 2002 Landsat images.

  Species specific basal area and biomass are recorded as a fraction of totals.
}

\usage{
data(BEF.dat)
}
 
\format{
 A data frame containing 437 rows and 208 columns.
}

\source{
  BEF.dat inventory data provided by:
  
  Marie-Louise Smith
  USDA Forest Service
  Northeastern Research Station
  \email{marielouisesmith@fs.fed.us}

  Additional variables provided by:
  
  Andrew Lister
  USDA Forest Service
  Northeastern Research Station
  \email{alister@fs.fed.us}
}


\keyword{datasets}
