\name{BEF}
\docType{data}

\alias{BEF}


\title{Bartlett Experimental Forest inventory data}

\description{
  Data generated in long-term research studies on the Bartlett
  Experimental Forest (BEF), Bartlett, NH funded by the U.S. Department
  of Agriculture, Forest Service, Northeastern Research Station.

  This dataset holds 2002 forest inventory data for 437 points
  on the BEF. Variables include total basal area, basal area of eastern
  hemlock, basal area of American beech, slope, elevation, and tasseled cap brightness (TC1),
  greenness (TC2), and wetness (TC3) components from spring, summer,
  and fall 2002 Landsat images.
}

\usage{
data(BEF)
}

\format{
 A data frame containing 437 rows and 16 columns.
}

\source{
  BEF inventory data provided by:
  
  Marie-Louise Smith
  USDA Forest Service
  Northeastern Research Station
  \email{marielouisesmith@fs.fed.us}

  Additional variables provided by:
  
  Andrew Lister
  USDA Forest Service
  Northeastern Research Station
  \email{alister@fs.fed.us}
}


\keyword{datasets}

