\name{FBC07.dat}
\docType{data}

\alias{FBC07.dat}


\title{Synthetic multivariate data with spatial and non-spatial variance
structures}

\description{
The synthetic dataset describes a stationary and isotropic bivariate process. Please refer to the vignette Section 4.2 for specifics.

}

\usage{
data(FBC07.dat)
}
 
\format{
 A data frame of 250 rows and 4 columns. Columns 1 and 2 are
 coordinates and columns 3 and 4 are response variables.
}

\source{
  Finley A.O., S. Banerjee, and B.P. Carlin (2007) spBayes: R package for Univariate and Multivariate
  Hierarchical Point-referenced Spatial Models. Journal of Statistical Software.
}

\keyword{datasets}

