\name{NETemp.dat}
\docType{data}

\alias{NETemp.dat}


\title{Monthly weather station temperature data across the Northeastern US}

\description{
  Monthly temperature data (Celsius) recorded across the Northeastern US
starting in January 2000. Station UTM coordinates and elevation are also included.

}

\usage{
data(NETemp.dat)
}
 
\format{
  A data frame containing 356 rows (weather stations) and 132 columns. 
}

\keyword{datasets}
