\name{WEF.dat}
\docType{data}

\alias{WEF.dat}


\title{Western Experimental Forest inventory data}

\description{
  
  Data generated as part of a long-term research study on an experimental forest
  in central Oregon.  This dataset holds the coordinates for all trees in the experimental forest. The typical stem measurements are recorded for each tree. Crown radius was measured at the cardinal directions for a subset of trees. Mean crown radius was calculated for all trees using a simple relationship between DBH, Height, and observed crown dimension.

}

\usage{
data(WEF.dat)
}
 
\format{
 A data frame containing 2422 rows and 15 columns.
}

\keyword{datasets}
